move prom-metrics under api
